+++ components/src/main/java/org/patternfly/component/menu/MenuGroup.java
@@ -71,6 +71,10 @@ public class MenuGroup extends MenuSubComponent<HTMLElement, MenuGroup> implemen
         return this;
     }
 
+    public MenuGroup addSearch(MenuSearch search) {
+        return add(search);
+    }
+
     // ------------------------------------------------------ builder
 
     @Override